+++ crates/tytanic/src/cli/commands/new.rs
@@ -80,7 +80,7 @@ pub fn run(ctx: &mut Context, args: &Args) -> eyre::Result<()> {
         Kind::CompileOnly => None,
         Kind::Ephemeral => Some(Reference::Ephemeral(source.into())),
         Kind::Persistent => {
-            let world = ctx.world()?;
+            let world = ctx.world(&args.compile)?;
             let path = project.paths().template();
 
             let path = path