+++ crates/tytanic/src/kit.rs
@@ -5,15 +5,19 @@ use typst_kit::download::Downloader;
 use typst_kit::fonts::{FontSearcher, Fonts};
 use typst_kit::package::PackageStorage;
 
-use crate::cli::commands::{FontOptions, PackageOptions, TypstOptions};
+use crate::cli::commands::{CompileOptions, FontOptions, PackageOptions, TypstOptions};
 use crate::world::SystemWorld;
 
-pub fn world(project_root: PathBuf, typst_options: &TypstOptions) -> eyre::Result<SystemWorld> {
+pub fn world(
+    project_root: PathBuf,
+    typst_options: &TypstOptions,
+    compile_options: &CompileOptions,
+) -> eyre::Result<SystemWorld> {
     let world = SystemWorld::new(
         project_root,
         fonts_from_args(&typst_options.font),
         package_storage_from_args(&typst_options.package),
-        typst_options.creation_timestamp,
+        compile_options.timestamp,
     )?;
 
     Ok(world)